requirements.txt
matplotlib==3.8.3
numpy==1.26.4
Pillow==10.2.0
Requests==2.31.0
streamlit==1.31.1
